#172: Replaced Invocation class with TransactionBuilder class.
