--- contract/src/main/java/io/neow3j/contract/Invocation.java
@@ -1,635 +1,0 @@
-package io.neow3j.contract;
-
-import static java.util.Arrays.asList;
-
-import io.neow3j.constants.InteropServiceCode;
-import io.neow3j.constants.NeoConstants;
-import io.neow3j.constants.OpCode;
-import io.neow3j.contract.exceptions.InvocationConfigurationException;
-import io.neow3j.crypto.ECKeyPair;
-import io.neow3j.crypto.ECKeyPair.ECPublicKey;
-import io.neow3j.crypto.Sign;
-import io.neow3j.crypto.Sign.SignatureData;
-import io.neow3j.io.IOUtils;
-import io.neow3j.protocol.Neow3j;
-import io.neow3j.protocol.core.methods.response.NeoInvokeFunction;
-import io.neow3j.protocol.core.methods.response.NeoInvokeScript;
-import io.neow3j.protocol.core.methods.response.NeoSendRawTransaction;
-import io.neow3j.transaction.Signer;
-import io.neow3j.transaction.Transaction;
-import io.neow3j.transaction.TransactionAttribute;
-import io.neow3j.transaction.VerificationScript;
-import io.neow3j.transaction.Witness;
-import io.neow3j.transaction.WitnessScope;
-import io.neow3j.utils.Numeric;
-import io.neow3j.wallet.Account;
-import io.neow3j.wallet.Wallet;
-import java.io.IOException;
-import java.math.BigInteger;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
-import java.util.function.BiConsumer;
-import java.util.function.Supplier;
-import java.util.stream.Collectors;
-
-/**
- * Used to invoke Neo VM scripts and contract functions. Uses the {@link Invocation.Builder} to
- * configure an invocation and do calls via the {@code invokescript} and {@code invokefunction} RPC.
- * When building the {@code Invocation}, a transaction is created that can be signed and sent to the
- * Neo node.
- */
-public class Invocation {
-
-    protected Neow3j neow;
-    protected Wallet wallet;
-    protected Transaction transaction;
-
-    protected Invocation(Builder builder) {
-        this.neow = builder.neow;
-        this.wallet = builder.wallet;
-        this.transaction = builder.transaction;
-    }
-
-    /**
-     * Sends this invocation transaction to the neo-node via the `sendrawtransaction` RPC.
-     *
-     * @return the Neo node's response.
-     * @throws IOException                      if a problem in communicating with the Neo node
-     *                                          occurs.
-     * @throws InvocationConfigurationException if signatures are missing for one or more signers of
-     *                                          the transaction.
-     */
-    public NeoSendRawTransaction send() throws IOException {
-        List<ScriptHash> witnesses = this.transaction.getWitnesses().stream()
-                .map(Witness::getScriptHash).collect(Collectors.toList());
-
-        for (Signer signer : this.transaction.getSigners()) {
-            if (!witnesses.contains(signer.getScriptHash())) {
-                throw new InvocationConfigurationException("The transaction does not have a "
-                        + "signature for each of its signers.");
-            }
-        }
-        String hex = Numeric.toHexStringNoPrefix(this.transaction.toArray());
-        return neow.sendRawTransaction(hex).send();
-    }
-
-    /**
-     * Creates signatures for every signer of the invocation transaction and adds them to the
-     * transaction as witnesses.
-     * <p>
-     * For each signer set on the transaction a corresponding account with an EC key pair must exist
-     * in the wallet set on the builder.
-     *
-     * @return this.
-     */
-    public Invocation sign() {
-        byte[] txBytes = getTransactionForSigning();
-        this.transaction.getSigners().forEach(signer -> {
-            if (!this.wallet.holdsAccount(signer.getScriptHash())) {
-                throw new InvocationConfigurationException("Can't create transaction "
-                        + "signature. Wallet does not contain the signer account with script "
-                        + "hash " + signer.getScriptHash());
-            } else {
-                Account signerAcc = this.wallet.getAccount(signer.getScriptHash());
-                if (signerAcc.isMultiSig()) {
-                    signWithMultiSigAccount(txBytes, signerAcc);
-                } else {
-                    signWithNormalAccount(txBytes, signerAcc);
-                }
-            }
-        });
-        return this;
-    }
-
-    private void signWithNormalAccount(byte[] txBytes, Account acc) {
-        ECKeyPair keyPair = acc.getECKeyPair();
-        if (keyPair == null) {
-            throw new InvocationConfigurationException("Can't create transaction signature "
-                    + "because account with script " + acc.getScriptHash() + " doesn't hold a "
-                    + "private key.");
-        }
-        this.transaction.addWitness(Witness.createWitness(txBytes, keyPair));
-    }
-
-    private void signWithMultiSigAccount(byte[] txBytes, Account signerAcc) {
-        List<SignatureData> sigs = new ArrayList<>();
-        VerificationScript multiSigVerifScript = signerAcc.getVerificationScript();
-        for (ECPublicKey pubKey : multiSigVerifScript.getPublicKeys()) {
-            ScriptHash accScriptHash = ScriptHash.fromPublicKey(pubKey.getEncoded(true));
-            Account a = this.wallet.getAccount(accScriptHash);
-            if (a == null || a.getECKeyPair() == null) {
-                continue;
-            }
-            sigs.add(Sign.signMessage(txBytes, a.getECKeyPair()));
-        }
-        int m = multiSigVerifScript.getSigningThreshold();
-        if (sigs.size() < m) {
-            throw new InvocationConfigurationException("Can't create transaction "
-                    + "signature. Wallet does not contain enough accounts (with decrypted "
-                    + "private keys) that are part of the multi-sig account with script "
-                    + "hash " + signerAcc.getScriptHash() + ".");
-        }
-        this.transaction.addWitness(Witness.createMultiSigWitness(sigs,
-                multiSigVerifScript));
-    }
-
-    /**
-     * Gets the invocation transaction for signing it.
-     *
-     * @return the transaction data ready for creating a signature.
-     */
-    public byte[] getTransactionForSigning() {
-        return this.transaction.getHashData();
-    }
-
-    /**
-     * Gets the invocation transaction.
-     *
-     * @return the transaction.
-     */
-    public Transaction getTransaction() {
-        return this.transaction;
-    }
-
-    /**
-     * Adds the given witnesses to the invocation transaction.
-     * <p>
-     * Use this method if you can't use the automatic signing method {@link Invocation#sign()},
-     * e.g., because the configured wallet does not contain all accounts needed for signing.
-     *
-     * @param witnesses The witnesses to add.
-     */
-    public void addWitnesses(Witness... witnesses) {
-        for (Witness witness : witnesses) {
-            this.transaction.addWitness(witness);
-        }
-    }
-
-    /**
-     * Checks if the sender account of this invocation can cover the network and system fees. If
-     * not, executes the given consumer supplying it with the required fee and the sender's GAS
-     * balance.
-     *
-     * @return this.
-     * @throws IOException if something goes wrong in the communication with the neo-node.
-     */
-    public Invocation doIfSenderCannotCoverFees(BiConsumer<BigInteger, BigInteger> consumer)
-            throws IOException {
-        BigInteger fees = BigInteger.valueOf(
-                this.transaction.getSystemFee() + this.transaction.getNetworkFee());
-        BigInteger senderGasBalance = new GasToken(this.neow)
-                .getBalanceOf(this.transaction.getSender());
-        if (fees.compareTo(senderGasBalance) > 0) {
-            consumer.accept(fees, senderGasBalance);
-        }
-        return this;
-    }
-
-    /**
-     * Checks if the sender account of this invocation can cover the network and system fees. If
-     * not, otherwise throw an exception created by the provided supplier.
-     *
-     * @return this.
-     * @throws IOException if something goes wrong in the communication with the neo-node.
-     */
-    public <T extends Throwable> Invocation throwIfSenderCannotCoverFees(
-            Supplier<? extends T> exceptionSupplier) throws IOException, T {
-
-        if (!canSenderCoverFees()) {
-            throw exceptionSupplier.get();
-        }
-        return this;
-    }
-
-    private boolean canSenderCoverFees() throws IOException {
-        BigInteger fees = BigInteger.valueOf(
-                this.transaction.getSystemFee() + this.transaction.getNetworkFee());
-        BigInteger senderGasBalance = new GasToken(this.neow)
-                .getBalanceOf(this.transaction.getSender());
-        return fees.compareTo(senderGasBalance) < 0;
-    }
-
-    public static class Builder {
-
-        private long additionalNetworkFee;
-        protected Neow3j neow;
-        protected Wallet wallet;
-        private String contractFunction;
-        private List<ContractParameter> contractParams;
-        private ScriptHash contract;
-        protected Transaction.Builder txBuilder;
-        protected Transaction transaction;
-        protected boolean failOnFalse;
-        private ScriptHash sender;
-
-        /**
-         * Constructs a builder for configuring a contract invocation.
-         *
-         * @param neow The {@link Neow3j} instance to use for JSON-RPC calls.
-         */
-        public Builder(Neow3j neow) {
-            if (neow == null) {
-                throw new IllegalArgumentException("Neow3j instance must not be null.");
-            }
-            this.txBuilder = new Transaction.Builder();
-            this.contractParams = new ArrayList<>();
-            this.neow = neow;
-        }
-
-        /**
-         * Configures the invocation with the given attributes.
-         *
-         * @param attributes The attributes.
-         * @return this.
-         */
-        public Builder withAttributes(TransactionAttribute... attributes) {
-            this.txBuilder.attributes(attributes);
-            return this;
-        }
-
-        /**
-         * Configures the invocation such that it is valid until the given block number.
-         * <p>
-         * By default it is set to the maximum.
-         *
-         * @param blockNr The black number
-         * @return this.
-         * @see Transaction.Builder#validUntilBlock(long)
-         */
-        public Builder withValidUntilBlock(long blockNr) {
-            this.txBuilder.validUntilBlock(blockNr);
-            return this;
-        }
-
-        /**
-         * Configures the invocation with the given signers.
-         * <p>
-         * If no sender is specified explicitly, the first signer parameter is used as the
-         * transaction sender.
-         *
-         * @param signers The signers.
-         * @return this.
-         */
-        public Builder withSigners(Signer... signers) {
-            this.txBuilder.signers(signers);
-            return this;
-        }
-
-        /**
-         * Configures the invocation with the given nonce.
-         * <p>
-         * By default the nonce is set to a random value.
-         *
-         * @param nonce The nonce.
-         * @return this.
-         * @see Transaction.Builder#nonce(Long)
-         */
-        public Builder withNonce(long nonce) {
-            this.txBuilder.nonce(nonce);
-            return this;
-        }
-
-        /**
-         * Configures the invocation with an additional network fee.
-         * <p>
-         * The basic network fee required to send this invocation is added automatically.
-         *
-         * @param fee The additional network fee in fractions of GAS.
-         * @return this.
-         */
-        public Builder withAdditionalNetworkFee(long fee) {
-            this.additionalNetworkFee = fee;
-            return this;
-        }
-
-        /**
-         * Configures the invocation to use the given wallet.
-         * <p>
-         * The wallet's default account is used as the transaction sender if no signers and no
-         * sender is specified explicitly.
-         *
-         * @param wallet The wallet.
-         * @return this.
-         */
-        public Builder withWallet(Wallet wallet) {
-            this.wallet = wallet;
-            return this;
-        }
-
-        /**
-         * Configures the invocation to use the given sender.
-         *
-         * @param sender the sender account's script hash.
-         * @return this.
-         */
-        public Builder withSender(ScriptHash sender) {
-            this.sender = sender;
-            return this;
-        }
-
-        /**
-         * Configures the invocation to call the function (configured via {@link
-         * Invocation.Builder#withFunction(String)}) with the provided parameters. The order of the
-         * parameters is relevant.
-         *
-         * @param parameters The contract parameters.
-         * @return this.
-         */
-        public Builder withParameters(ContractParameter... parameters) {
-            this.contractParams.addAll(asList(parameters));
-            return this;
-        }
-
-        /**
-         * Configures the invocation to call the provided function on the contract configured via
-         * {@link Invocation.Builder#withContract(ScriptHash)}.
-         *
-         * @param function The contract function to call.
-         * @return this.
-         */
-        public Builder withFunction(String function) {
-            this.contractFunction = function;
-            return this;
-        }
-
-        /**
-         * Configures the invocation to call the given contract.
-         *
-         * @param contract The script hash of the contract to call.
-         * @return this.
-         */
-        public Builder withContract(ScriptHash contract) {
-            this.contract = contract;
-            return this;
-        }
-
-        /**
-         * Configures the invocation to run the given script.
-         *
-         * @param script The script to invoke.
-         * @return this.
-         */
-        public Builder withScript(byte[] script) {
-            this.txBuilder.script(script);
-            return this;
-        }
-
-        /**
-         * Configures the invocation such that it fails (NeoVM exits with state FAULT) if the return
-         * value of the invocation is "False".
-         *
-         * @return this
-         */
-        public Builder failOnFalse() {
-            this.failOnFalse = true;
-            return this;
-        }
-
-        /**
-         * Makes an {@code invokescript} call to the neo-node with the invocation in its current
-         * configuration. No changes are made to the blockchain state.
-         * <p>
-         * Make sure to add all necessary signers to the builder before making this call. They are
-         * required for a successful {@code invokescript} call.
-         *
-         * @return the call's response.
-         * @throws IOException if something goes wrong when communicating with the neo-node.
-         */
-        public NeoInvokeScript invokeScript() throws IOException {
-            if (this.txBuilder.getScript() == null || this.txBuilder.getScript().length == 0) {
-                throw new InvocationConfigurationException("Cannot make an 'invokescript' call "
-                        + "without the script being configured.");
-            }
-            // The list of signers is required for `invokescript` calls that will hit a
-            // ChecekWitness check in the smart contract. We add the signers even if that
-            // is not the case because we cannot know if the invoked script needs it or not and it
-            // doesn't lead to failures if we add them in any case.
-            Signer[] signers = this.txBuilder.getSigners().toArray(new Signer[0]);
-            String script = Numeric.toHexStringNoPrefix(this.txBuilder.getScript());
-            return neow.invokeScript(script, signers).send();
-        }
-
-        /**
-         * Makes an {@code invokefunction} call to the neo-node with the invocation in its current
-         * configuration. No changes are made to the blockchain state.
-         * <p>
-         * Make sure to add all necessary signers to the builder before making this call. They are
-         * required for a successful {@code invokefunction} call.
-         *
-         * @return the call's response.
-         * @throws IOException if something goes wrong when communicating with the neo-node.
-         */
-        public NeoInvokeFunction invokeFunction() throws IOException {
-            if (this.contract == null) {
-                throw new InvocationConfigurationException("Cannot make an 'invokefunction' call "
-                        + "without having configured the contract to call.");
-            }
-            if (this.contractFunction == null) {
-                throw new InvocationConfigurationException("Cannot make an 'invokefunction' call "
-                        + "without having a configured function to call.");
-            }
-
-            // The list of signers is required for `invokefunction` calls that will hit a
-            // CheckWitness check in the smart contract. We add the signers even if that is not the
-            // case because we cannot know if the invoked function needs it or not and it doesn't
-            // lead to failures if we add them in any case.
-            Signer[] signers = this.txBuilder.getSigners().toArray(new Signer[0]);
-            if (this.contractParams.isEmpty()) {
-                return neow.invokeFunction(this.contract.toString(), this.contractFunction, null,
-                        signers).send();
-            }
-            return neow.invokeFunction(this.contract.toString(), this.contractFunction,
-                    this.contractParams, signers).send();
-        }
-
-        /**
-         * Builds the invocation, enforces correct configuration, fetches the system fee and
-         * calculates the network fee.
-         *
-         * @return the {@code Invocation} ready for signing and sending.
-         * @throws IOException if something goes wrong when communicating with the neo-node.
-         */
-        public Invocation build() throws IOException {
-            if (this.wallet == null) {
-                throw new InvocationConfigurationException("Cannot build a transaction without a "
-                        + "wallet.");
-            }
-            if (this.txBuilder.getValidUntilBlock() == null) {
-                // If validUntilBlock is not set explicitly set, then set it to the current max.
-                // It can happen that the neo-node refuses the valid until block because of
-                // it being over the max. Therefore, we decrement it by 1, to make sure that
-                // the node doesn't reject the transaction.
-                this.txBuilder.validUntilBlock(
-                        fetchCurrentBlockNr() + NeoConstants.MAX_VALID_UNTIL_BLOCK_INCREMENT - 1);
-            }
-            prepareTransactionSigners();
-            if (this.txBuilder.getScript() == null || this.txBuilder.getScript().length == 0) {
-                // The builder was not configured with a script. Therefore, try to construct one
-                // from the contract, function, and parameters.
-                this.txBuilder.script(buildScript());
-            }
-            this.txBuilder.systemFee(getSystemFeeForScript());
-            this.txBuilder.networkFee(calcNetworkFee() + this.additionalNetworkFee);
-            this.transaction = this.txBuilder.build();
-            return new Invocation(this);
-        }
-
-        // Prepares the list of signers such that there is at least one signer that will be used
-        // to cover the transaction fees.
-        private void prepareTransactionSigners() {
-            if (this.sender != null) {
-                if (hasSignerWithFeeOnlyScope()) {
-                    throw new InvocationConfigurationException("The list of transaction signers "
-                            + "contained a signer with the fee-only scope and at the same time a "
-                            + "sender was set. Either a fee-only signer, or a sender is allowed, "
-                            + "but not both at the same time.");
-                }
-                Optional<Signer> senderInSignersOpt = this.txBuilder.getSigners().stream()
-                        .filter(s -> s.getScriptHash().equals(this.sender))
-                        .findFirst();
-                if (senderInSignersOpt.isPresent()) {
-                    // If the sender is already in the signers list move it to the first position in
-                    // the signer list, but don't change the scope (can be more than feeOnly).
-                    this.txBuilder.getSigners().remove(senderInSignersOpt.get());
-                    this.txBuilder.getSigners().add(0, senderInSignersOpt.get());
-                } else {
-                    // Add the sender to the first position in the signer list.
-                    this.txBuilder.getSigners().add(0, Signer.feeOnly(this.sender));
-                }
-            } else if (this.txBuilder.getSigners().isEmpty()) {
-                // If no signer or sender is present, add the default account as a signer with
-                // the feeOnly witness scope. It will be used to cover the fees.
-                this.txBuilder.signers(Signer.feeOnly(
-                        this.wallet.getDefaultAccount().getScriptHash()));
-            }
-        }
-
-        // Checks if there is a signer (excluding the sender) with a feeOnly scope.
-        private boolean hasSignerWithFeeOnlyScope() {
-            return this.txBuilder.getSigners().stream()
-                    .anyMatch(s -> s.getScopes().contains(WitnessScope.FEE_ONLY)
-                            && !s.getScriptHash().equals(this.sender));
-        }
-
-        // Tries to build a script from the properties contract, function and parameters
-        // configured on this builder.
-        private byte[] buildScript() {
-            if (this.contract == null) {
-                throw new InvocationConfigurationException("The invocation doesn't have a script to"
-                        + " invoke and can't generate one from because the contract to invoke was "
-                        + "not set.");
-            }
-            if (this.contractFunction == null) {
-                throw new InvocationConfigurationException("The invocation is configured to call a "
-                        + "contract but is missing a specific function to call.");
-            }
-            ScriptBuilder b = new ScriptBuilder()
-                    .contractCall(this.contract, this.contractFunction, this.contractParams);
-            if (this.failOnFalse) {
-                b.opCode(OpCode.ASSERT);
-            }
-            return b.toArray();
-        }
-
-        private long fetchCurrentBlockNr() throws IOException {
-            return neow.getBlockCount().send().getBlockIndex().longValue();
-        }
-
-        /*
-         * Fetches the GAS consumed by this invocation. It does this by making an RPC call to the
-         * Neo node. The returned GAS amount is in fractions of GAS (10^-8).
-         */
-        private long getSystemFeeForScript() throws IOException {
-            // The signers are required for `invokescript` calls that will hit a CheckWitness
-            // check in the smart contract.
-            Signer[] signers = this.txBuilder.getSigners().toArray(new Signer[0]);
-            String script = Numeric.toHexStringNoPrefix(this.txBuilder.getScript());
-            NeoInvokeScript response = neow.invokeScript(script, signers).send();
-            // The GAS amount is returned in fractions (10^8)
-            return Long.parseLong(response.getInvocationResult().getGasConsumed());
-        }
-
-        /*
-         * Calculates the necessary network fee for the transaction being build in this builder.
-         * The fee consists of the cost per transaction byte and the cost for signature
-         * verification. Since the transaction is not signed yet, the calculation works with
-         * expected signatures. This information is derived from the verification scripts of all
-         * signers added to the transaction.
-         */
-        private long calcNetworkFee() {
-            List<Account> cosigAccs = getSignerAccounts();
-
-            // Base transaction size
-            int size = Transaction.HEADER_SIZE // constant header size
-                    + IOUtils.getVarSize(this.txBuilder.getSigners())
-                    + IOUtils.getVarSize(this.txBuilder.getAttributes()) // attributes
-                    + IOUtils.getVarSize(this.txBuilder.getScript()) // script
-                    + IOUtils.getVarSize(cosigAccs.size()); // varInt for all necessary witnesses
-
-            // Calculate fee for witness verification and collect size of witnesses.
-            int execFee = 0;
-            for (Account acc : cosigAccs) {
-                if (acc.isMultiSig()) {
-                    size += calcSizeForMultiSigWitness(acc.getVerificationScript());
-                    execFee += calcExecutionFeeForMultiSigWitness(acc.getVerificationScript());
-                } else {
-                    size += calcSizeForSingleSigWitness(acc.getVerificationScript());
-                    execFee += calcExecutionFeeForSingleSigWitness();
-                }
-            }
-            return execFee + size * NeoConstants.GAS_PER_BYTE;
-        }
-
-        /**
-         * Gets the signer accounts held in the wallet.
-         *
-         * @return a list containing the signer accounts
-         */
-        private List<Account> getSignerAccounts() {
-            List<Account> accounts = new ArrayList<>();
-            txBuilder.getSigners().forEach(signer -> {
-                if (this.wallet.holdsAccount(signer.getScriptHash())) {
-                    accounts.add(this.wallet.getAccount(signer.getScriptHash()));
-                } else {
-                    throw new InvocationConfigurationException("Wallet does not contain the "
-                            + "account for signer with script hash " + signer.getScriptHash());
-                }
-            });
-            return accounts;
-        }
-
-        private long calcSizeForSingleSigWitness(VerificationScript verifScript) {
-            return NeoConstants.SERIALIZED_INVOCATION_SCRIPT_SIZE + verifScript.getSize();
-        }
-
-        private long calcExecutionFeeForSingleSigWitness() {
-            return OpCode.PUSHDATA1.getPrice() // Push invocation script
-                    + OpCode.PUSHDATA1.getPrice() // Push verification script
-                    // Push null because we don't want to verify a particular message but the
-                    // transaction itself.
-                    + OpCode.PUSHNULL.getPrice()
-                    + InteropServiceCode.NEO_CRYPTO_VERIFYWITHECDSASECP256R1.getPrice();
-        }
-
-        private long calcSizeForMultiSigWitness(VerificationScript verifScript) {
-            int m = verifScript.getSigningThreshold();
-            int sizeInvocScript = NeoConstants.INVOCATION_SCRIPT_SIZE * m;
-            return IOUtils.getVarSize(sizeInvocScript) + sizeInvocScript + verifScript.getSize();
-        }
-
-        private long calcExecutionFeeForMultiSigWitness(VerificationScript verifScript) {
-            int m = verifScript.getSigningThreshold();
-            int n = verifScript.getNrOfAccounts();
-
-            return OpCode.PUSHDATA1.getPrice() * m
-                    + OpCode.get(new ScriptBuilder().pushInteger(m).toArray()[0]).getPrice()
-                    + OpCode.PUSHDATA1.getPrice() * n
-                    + OpCode.get(new ScriptBuilder().pushInteger(n).toArray()[0]).getPrice()
-                    // Push null because we don't want to verify a particular message but the
-                    // transaction itself.
-                    + OpCode.PUSHNULL.getPrice()
-                    + InteropServiceCode.NEO_CRYPTO_CHECKMULTISIGWITHECDSASECP256R1.getPrice(n);
-        }
-    }
-}
